Move `hit` and `lint` to `backend`
